Implement authorization using policy

+++ app/Http/Controllers/QuestionsController.php
@@ -8,6 +8,10 @@ use Illuminate\Http\Request;
 
 class QuestionsController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
     /**
      * Display a listing of the resource.
      *
@@ -17,7 +21,7 @@ class QuestionsController extends Controller
     {
         // \DB::enableQueryLog();
         $questions = Question::with('user')->latest()->paginate(5);
-       return view('questions.index', compact('questions'));
+        return view('questions.index', compact('questions'));
         // dd(\DB::getQueryLog());
     }
 
@@ -29,7 +33,7 @@ class QuestionsController extends Controller
     public function create()
     {
         $question = new Question();
-        return view('questions.create',compact('question'));
+        return view('questions.create', compact('question'));
     }
 
     /**
@@ -40,8 +44,8 @@ class QuestionsController extends Controller
      */
     public function store(AskQuestionRequest $request)
     {
-        $request->user()->questions()->create($request->only('title','body'));
-        return redirect()->route('questions.index')->with('success','Your question has been submitted');
+        $request->user()->questions()->create($request->only('title', 'body'));
+        return redirect()->route('questions.index')->with('success', 'Your question has been submitted');
     }
 
     /**
@@ -53,7 +57,7 @@ class QuestionsController extends Controller
     public function show(Question $question)
     {
         $question->increment('views');
-        return view('questions.show',compact('question'));
+        return view('questions.show', compact('question'));
     }
 
     /**
@@ -64,7 +68,7 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
-        return view('questions.edit',compact('question'));
+        return view('questions.edit', compact('question'));
     }
 
     /**
@@ -76,8 +80,9 @@ class QuestionsController extends Controller
      */
     public function update(Request $request, Question $question)
     {
-        $question->update($request->only('title','body'));
-        return redirect()->route('questions.index')->with('success','Your question has been updated');
+        $this->authorize('update', $question);
+        $question->update($request->only('title', 'body'));
+        return redirect()->route('questions.index')->with('success', 'Your question has been updated');
     }
 
     /**
@@ -88,7 +93,8 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        $this->authorize('delete', $question);
         $question->delete();
-        return redirect()->route('questions.index')->with('success','Your question has been deleted');
+        return redirect()->route('questions.index')->with('success', 'Your question has been deleted');
     }
 }
